+++ packages/runtime-html/src/resources/custom-attribute.ts
@@ -4,7 +4,7 @@ import { Bindable } from '../bindable';
 import { Watch } from '../watch';
 import { getRef } from '../dom';
 import { DefinitionType } from './resources-shared';
-import { appendResourceKey, defineMetadata, getAnnotationKeyFor, getOwnMetadata, getResourceKeyFor, hasOwnMetadata } from '../shared';
+import { appendResourceKey, defineMetadata, getAnnotationKeyFor, getOwnMetadata, getResourceKeyFor, hasOwnMetadata } from '../utilities-metadata';
 import { isFunction, isString } from '../utilities';
 import { aliasRegistration, transientRegistration } from '../utilities-di';
