+++ packages/runtime-html/src/templating/view.ts
@@ -2,7 +2,7 @@ import { DI } from '@aurelia/kernel';
 import { Scope } from '@aurelia/runtime';
 import { CustomElementDefinition, defineElement, getElementDefinition } from '../resources/custom-element';
 import { Controller } from './controller';
-import { defineMetadata, getOwnMetadata, getResourceKeyFor, hasOwnMetadata } from '../shared';
+import { defineMetadata, getOwnMetadata, getResourceKeyFor, hasOwnMetadata } from '../utilities-metadata';
 import { isFunction, isString } from '../utilities';
 
 import type { Constructable, ConstructableClass, IContainer } from '@aurelia/kernel';
