+++ packages/runtime-html/src/watch.ts
@@ -2,7 +2,7 @@
 import { emptyArray } from '@aurelia/kernel';
 import { getAttributeDefinition, isAttributeType } from './resources/custom-attribute';
 import { getElementDefinition, isElementType } from './resources/custom-element';
-import { defineMetadata, getAnnotationKeyFor, getOwnMetadata } from './shared';
+import { defineMetadata, getAnnotationKeyFor, getOwnMetadata } from './utilities-metadata';
 import { isFunction } from './utilities';
 
 import type { Constructable } from '@aurelia/kernel';
